Implement AuthService as a subclass of BaseServicesClient.
Manages refresh_token, access_token, and access_token_expiration, with a fixed endpoint_base.

The authenticate() method calls the HTTP client to request a new token and sets a 5‑minute expiration.expiration of 5 min

The is_access_token_valid() check if the token is not none and if the expiration date is not exceeded

Introduce AuthResponse type to define the expected structure of authentication API responses.